Title: 시력 검사, Time: 175ms, Memory: 11MB, Status: Passed - CodeTree

+++ 240826/시력 검사/eye-test.java
@@ -6,8 +6,8 @@ public class Main {
 
         Scanner sc = new Scanner(System.in);
 
-        double a = sc.nextInt();
-        double b = sc.nextInt();
+        double a = sc.nextDouble();
+        double b = sc.nextDouble();
